Recursion Implementation
Recursion did not allow the implementation of the designed solution for the conditional routing. I may have to restructure the components in order to use Context, rather than relying on localStorage to run the authentication check. If time does not permit it, I might just style the pages as best I can, so that at least that part of the spec is met. Need to familiarise myself more with react in order to properly implement conditionalrouting/rendering